Refactor DocumentExecuter to use "execution strategies"
