+++ src/GraphQL/Execution/DocumentExecuter.cs
@@ -1,18 +1,15 @@
 using System;
-using System.Collections;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using GraphQL.Execution;
 using GraphQL.Instrumentation;
-using GraphQL.Introspection;
 using GraphQL.Language.AST;
-using GraphQL.Resolvers;
 using GraphQL.Types;
 using GraphQL.Validation;
 using GraphQL.Validation.Complexity;
+using static GraphQL.Execution.ExecutionHelper;
 using ExecutionContext = GraphQL.Execution.ExecutionContext;
 using Field = GraphQL.Language.AST.Field;
 
@@ -79,44 +76,12 @@ namespace GraphQL
 
         public Task<ExecutionResult> ExecuteAsync(Action<ExecutionOptions> configure)
         {
+            if (configure == null)
+                throw new ArgumentNullException(nameof(configure));
+
             var options = new ExecutionOptions();
             configure(options);
             return ExecuteAsync(options);
-        }
-
-        protected virtual ExecutionResult InitializeResult(ExecutionOptions config)
-        {
-            var result = new ExecutionResult { Query = config.Query, ExposeExceptions = config.ExposeExceptions };
-            return result;
-        }
-
-        protected virtual async Task OnExecution(ExecutionOptions config, ExecutionContext context, ExecutionResult result)
-        {
-            var task = ExecuteOperationAsync(context).ConfigureAwait(false);
-
-            foreach (var listener in config.Listeners)
-            {
-                await listener.BeforeExecutionAwaitedAsync(config.UserContext, config.CancellationToken).ConfigureAwait(false);
-            }
-
-            result.Data = await task;
-        }
-
-        protected virtual void OnValidationError(IValidationResult validationResult, ExecutionResult result)
-        {
-            result.Data = null;
-            result.Errors = validationResult.Errors;
-        }
-
-        protected virtual void OnError(Exception exception, ExecutionResult result)
-        {
-            if (result.Errors == null)
-            {
-                result.Errors = new ExecutionErrors();
-            }
-
-            result.Data = null;
-            result.Errors.Add(new ExecutionError(exception.Message, exception));
         }
 
         private void ValidateOptions(ExecutionOptions options)
@@ -132,43 +97,44 @@ namespace GraphQL
             }
         }
 
-        public async Task<ExecutionResult> ExecuteAsync(ExecutionOptions config)
+        public async Task<ExecutionResult> ExecuteAsync(ExecutionOptions options)
         {
-            var metrics = new Metrics(config.EnableMetrics);
-            metrics.Start(config.OperationName);
+            if (options == null)
+                throw new ArgumentNullException(nameof(options));
 
-            config.Schema.FieldNameConverter = config.FieldNameConverter;
+            var metrics = new Metrics(options.EnableMetrics);
+            metrics.Start(options.OperationName);
 
-            var result = InitializeResult(config);
+            options.Schema.FieldNameConverter = options.FieldNameConverter;
+
+            ExecutionResult result = null;
+
             try
             {
-                ValidateOptions(config);
+                ValidateOptions(options);
 
-                if (!config.Schema.Initialized)
+                if (!options.Schema.Initialized)
                 {
                     using (metrics.Subject("schema", "Initializing schema"))
                     {
-                        if (config.SetFieldMiddleware)
+                        if (options.SetFieldMiddleware)
                         {
-                            config.FieldMiddleware.ApplyTo(config.Schema);
+                            options.FieldMiddleware.ApplyTo(options.Schema);
                         }
-                        config.Schema.Initialize();
+                        options.Schema.Initialize();
                     }
                 }
 
-                var document = config.Document;
+                var document = options.Document;
                 using (metrics.Subject("document", "Building document"))
                 {
                     if (document == null)
                     {
-                        document = _documentBuilder.Build(config.Query);
+                        document = _documentBuilder.Build(options.Query);
                     }
                 }
 
-                result.Document = document;
-
-                var operation = GetOperation(config.OperationName, document);
-                result.Operation = operation;
+                var operation = GetOperation(options.OperationName, document);
                 metrics.SetOperationName(operation?.Name);
 
                 if (operation == null)
@@ -180,83 +146,117 @@ namespace GraphQL
                 using (metrics.Subject("document", "Validating document"))
                 {
                     validationResult = _documentValidator.Validate(
-                        config.Query,
-                        config.Schema,
+                        options.Query,
+                        options.Schema,
                         document,
-                        config.ValidationRules,
-                        config.UserContext);
+                        options.ValidationRules,
+                        options.UserContext);
                 }
 
-                if (config.ComplexityConfiguration != null && validationResult.IsValid)
+                if (options.ComplexityConfiguration != null && validationResult.IsValid)
                 {
                     using (metrics.Subject("document", "Analyzing complexity"))
-                        _complexityAnalyzer.Validate(document, config.ComplexityConfiguration);
+                        _complexityAnalyzer.Validate(document, options.ComplexityConfiguration);
                 }
 
-                foreach (var listener in config.Listeners)
+                foreach (var listener in options.Listeners)
                 {
                     await listener.AfterValidationAsync(
-                            config.UserContext,
+                            options.UserContext,
                             validationResult,
-                            config.CancellationToken)
+                            options.CancellationToken)
                         .ConfigureAwait(false);
                 }
 
-                if (validationResult.IsValid)
+                if (!validationResult.IsValid)
                 {
-                    var context = BuildExecutionContext(
-                        config.Schema,
-                        config.Root,
-                        document,
-                        operation,
-                        config.Inputs,
-                        config.UserContext,
-                        config.CancellationToken,
-                        metrics,
-                        config.Listeners);
-
-                    if (context.Errors.Any())
+                    return new ExecutionResult()
                     {
-                        result.Errors = context.Errors;
-                        return result;
-                    }
-
-                    using (metrics.Subject("execution", "Executing operation"))
-                    {
-                        foreach (var listener in config.Listeners)
-                        {
-                            await listener.BeforeExecutionAsync(config.UserContext, config.CancellationToken).ConfigureAwait(false);
-                        }
-
-                        await OnExecution(config, context, result).ConfigureAwait(false);
-
-                        foreach (var listener in config.Listeners)
-                        {
-                            await listener.AfterExecutionAsync(config.UserContext, config.CancellationToken).ConfigureAwait(false);
-                        }
-                    }
-
-                    if (context.Errors.Any())
-                    {
-                        result.Errors = context.Errors;
-                    }
-                }
-                else
-                {
-                    OnValidationError(validationResult, result);
+                        Errors = validationResult.Errors
+                    };
                 }
 
-                return result;
+                var context = BuildExecutionContext(
+                    options.Schema,
+                    options.Root,
+                    document,
+                    operation,
+                    options.Inputs,
+                    options.UserContext,
+                    options.CancellationToken,
+                    metrics,
+                    options.Listeners);
+
+                if (context.Errors.Any())
+                {
+                    return new ExecutionResult()
+                    {
+                        Errors = context.Errors
+                    };
+                }
+
+                using (metrics.Subject("execution", "Executing operation"))
+                {
+                    foreach (var listener in context.Listeners)
+                    {
+                        await listener.BeforeExecutionAsync(context.UserContext, context.CancellationToken)
+                            .ConfigureAwait(false);
+                    }
+
+                    IExecutionStrategy executionStrategy = SelectExecutionStrategy(context);
+
+                    if (executionStrategy == null)
+                        throw new InvalidOperationException("Invalid ExecutionStrategy!");
+
+                    var rootType = GetOperationRootType(context.Document, context.Schema, context.Operation);
+                    var fields = CollectFields(
+                        context,
+                        rootType,
+                        context.Operation.SelectionSet,
+                        new Dictionary<string, Field>(),
+                        new List<string>());
+
+                    var task = executionStrategy.ExecuteAsync(context, rootType, context.RootValue, fields)
+                        .ConfigureAwait(false);
+
+                    foreach (var listener in context.Listeners)
+                    {
+                        await listener.BeforeExecutionAwaitedAsync(context.UserContext, context.CancellationToken)
+                            .ConfigureAwait(false);
+                    }
+
+                    result = await task;
+
+                    foreach (var listener in context.Listeners)
+                    {
+                        await listener.AfterExecutionAsync(context.UserContext, context.CancellationToken)
+                            .ConfigureAwait(false);
+                    }
+                }
+
+                if (context.Errors.Any())
+                {
+                    result.Errors = context.Errors;
+                }
             }
-            catch (Exception exc)
+            catch (Exception ex)
             {
-                OnError(exc, result);
-                return result;
+                result = new ExecutionResult
+                {
+                    Errors = new ExecutionErrors()
+                    {
+                        new ExecutionError(ex.Message, ex)
+                    }
+                };
             }
             finally
             {
+                result = result ?? new ExecutionResult();
+                result.ExposeExceptions = options.ExposeExceptions;
                 result.Perf = metrics.Finish()?.ToArray();
             }
+
+            return result;
         }
 
         public ExecutionContext BuildExecutionContext(
@@ -294,463 +294,6 @@ namespace GraphQL
                 : document.Operations.FirstOrDefault();
 
             return operation;
-        }
-
-        public Task<IDictionary<string, object>> ExecuteOperationAsync(ExecutionContext context)
-        {
-            var rootType = GetOperationRootType(context.Document, context.Schema, context.Operation);
-            var fields = CollectFields(
-                context,
-                rootType,
-                context.Operation.SelectionSet,
-                new Dictionary<string, Field>(),
-                new List<string>());
-
-            return ExecuteFieldsAsync(context, rootType, context.RootValue, fields, new string[0]);
-        }
-
-        public async Task<IDictionary<string, object>> ExecuteFieldsAsync(ExecutionContext context, IObjectGraphType rootType, object source, Dictionary<string, Field> fields, IEnumerable<string> path)
-        {
-            var data = new ConcurrentDictionary<string, object>();
-            var tasks = new List<Task>();
-
-            foreach (var fieldCollection in fields)
-            {
-                var currentPath = path.Concat(new[] { fieldCollection.Key });
-
-                var field = fieldCollection.Value;
-                var fieldType = GetFieldDefinition(context.Document, context.Schema, rootType, field);
-
-                if (fieldType?.Resolver == null || context.Operation.OperationType == OperationType.Mutation)
-                {
-                    await ExtractFieldAsync(context, rootType, source, field, fieldType, data, currentPath);
-                }
-                else
-                {
-                    var task = ExtractFieldAsync(context, rootType, source, field, fieldType, data, currentPath);
-                    tasks.Add(task);
-                }
-            }
-
-            foreach (var listener in context.Listeners)
-            {
-                await listener.BeforeResolveLevelAwaitedAsync(context.UserContext, context.CancellationToken)
-                    .ConfigureAwait(false);
-            }
-
-            if (tasks.Count > 0)
-            {
-                await Task.WhenAll(tasks.ToArray())
-                    .ConfigureAwait(false);
-            }
-
-            var ordered = new Dictionary<string, object>();
-
-            foreach (var fieldCollection in fields)
-            {
-                var name = fieldCollection.Key;
-
-                if (!data.ContainsKey(name))
-                {
-                    continue;
-                }
-
-                ordered.Add(name, data[name]);
-            }
-
-            return ordered;
-        }
-
-        private async Task ExtractFieldAsync(ExecutionContext context, IObjectGraphType rootType, object source,
-            Field field, FieldType fieldType, ConcurrentDictionary<string, object> data, IEnumerable<string> path)
-        {
-            context.CancellationToken.ThrowIfCancellationRequested();
-
-            var name = field.Alias ?? field.Name;
-
-            if (data.ContainsKey(name))
-            {
-                return;
-            }
-
-            if (!ShouldIncludeNode(context, field.Directives))
-            {
-                return;
-            }
-
-            var result = await ResolveFieldAsync(context, rootType, source, field, path)
-                .ConfigureAwait(false);
-
-            if (result.Skip)
-            {
-                return;
-            }
-
-            data.TryAdd(name, result.Value);
-        }
-
-        //private bool CanResolveFromData(Field field, FieldType type)
-        //{
-        //    if (field == null || type == null)
-        //    {
-        //        return false;
-        //    }
-
-        //    if (type.Arguments != null &&
-        //        type.Arguments.Any())
-        //    {
-        //        return false;
-        //    }
-
-        //    if (!(type.ResolvedType is ScalarGraphType))
-        //    {
-        //        return false;
-        //    }
-
-        //    if (type.ResolvedType is NonNullGraphType)
-        //    {
-        //        return false;
-        //    }
-
-        //    return true;
-        //}
-
-        ///// <summary>
-        ///// Resolve simple fields in a performant manner
-        ///// </summary>
-        //private static async Task<object> ResolveFieldFromDataAsync(ExecutionContext context, IObjectGraphType rootType, object source,
-        //    FieldType fieldType, Field field, IEnumerable<string> path)
-        //{
-        //    object result = null;
-
-        //    try
-        //    {
-        //        if (fieldType.Resolver != null)
-        //        {
-        //            var rfc = new ResolveFieldContext(context, field, fieldType, source, rootType, null, path);
-
-        //            result = fieldType.Resolver.Resolve(rfc);
-
-        //            result = await UnwrapResultAsync(result)
-        //                .ConfigureAwait(false);
-        //        }
-        //        else
-        //        {
-        //            result = NameFieldResolver.Resolve(source, field.Name);
-        //        }
-
-        //        if (result != null)
-        //        {
-        //            var scalarType = fieldType.ResolvedType as ScalarGraphType;
-
-        //            result = scalarType?.Serialize(result);
-        //        }
-        //    }
-        //    catch (Exception exc)
-        //    {
-        //        var error = new ExecutionError($"Error trying to resolve {field.Name}.", exc);
-        //        error.AddLocation(field, context.Document);
-        //        error.Path = path.ToList();
-        //        context.Errors.Add(error);
-
-        //        // If there was an exception, the value of result cannot be trusted
-        //        result = null;
-        //    }
-
-        //    return result;
-        //}
-
-        public async Task<ResolveFieldResult<object>> ResolveFieldAsync(ExecutionContext context, IObjectGraphType parentType, object source, Field field, IEnumerable<string> path)
-        {
-            context.CancellationToken.ThrowIfCancellationRequested();
-
-            var fieldPath = path?.ToList() ?? new List<string>();
-
-            var resolveResult = new ResolveFieldResult<object>
-            {
-                Skip = false
-            };
-
-            var fieldDefinition = GetFieldDefinition(context.Document, context.Schema, parentType, field);
-            if (fieldDefinition == null)
-            {
-                resolveResult.Skip = true;
-                return resolveResult;
-            }
-
-            var arguments = GetArgumentValues(context.Schema, fieldDefinition.Arguments, field.Arguments, context.Variables);
-
-            try
-            {
-                var resolveContext = new ResolveFieldContext
-                {
-                    FieldName = field.Name,
-                    FieldAst = field,
-                    FieldDefinition = fieldDefinition,
-                    ReturnType = fieldDefinition.ResolvedType,
-                    ParentType = parentType,
-                    Arguments = arguments,
-                    Source = source,
-                    Schema = context.Schema,
-                    Document = context.Document,
-                    Fragments = context.Fragments,
-                    RootValue = context.RootValue,
-                    UserContext = context.UserContext,
-                    Operation = context.Operation,
-                    Variables = context.Variables,
-                    CancellationToken = context.CancellationToken,
-                    Metrics = context.Metrics,
-                    Errors = context.Errors,
-                    Path = fieldPath
-                };
-
-                resolveContext.SubFields = SubFieldsFor(context, fieldDefinition.ResolvedType, field);
-
-                var resolver = fieldDefinition.Resolver ?? new NameFieldResolver();
-                var result = resolver.Resolve(resolveContext);
-
-                result = await UnwrapResultAsync(result)
-                    .ConfigureAwait(false);
-
-                resolveResult.Value = await CompleteValueAsync(context, parentType, fieldDefinition.ResolvedType, field, result, fieldPath)
-                    .ConfigureAwait(false);
-
-                return resolveResult;
-            }
-            catch (Exception exc)
-            {
-                return GenerateError(resolveResult, field, context, exc, path);
-            }
-        }
-
-        /// <summary>
-        /// Unwrap nested Tasks to get the result
-        /// </summary>
-        /// <param name="result"></param>
-        /// <returns></returns>
-        private static async Task<object> UnwrapResultAsync(object result)
-        {
-            while (result is Task task)
-            {
-                await task.ConfigureAwait(false);
-
-                // Most performant if available
-                if (task is Task<object> t)
-                {
-                    result = t.Result;
-                }
-                else
-                {
-                    result = ((dynamic)task).Result;
-                }
-            }
-
-            return result;
-        }
-
-        private IDictionary<string, Field> SubFieldsFor(ExecutionContext context, IGraphType fieldType, Field field)
-        {
-            if (!(fieldType is IObjectGraphType) || !field.SelectionSet.Selections.Any())
-            {
-                return null;
-            }
-
-            var subFields = new Dictionary<string, Field>();
-            var visitedFragments = new List<string>();
-            var fields = CollectFields(context, fieldType, field.SelectionSet, subFields, visitedFragments);
-            return fields;
-        }
-
-        private ResolveFieldResult<object> GenerateError(
-            ResolveFieldResult<object> resolveResult,
-            Field field,
-            ExecutionContext context,
-            Exception exc,
-            IEnumerable<string> path)
-        {
-            var error = new ExecutionError("Error trying to resolve {0}.".ToFormat(field.Name), exc);
-            error.AddLocation(field, context.Document);
-            error.Path = path;
-            context.Errors.Add(error);
-            resolveResult.Skip = false;
-            return resolveResult;
-        }
-
-        public async Task<object> CompleteValueAsync(ExecutionContext context, IObjectGraphType parentType, IGraphType fieldType, Field field, object result, IEnumerable<string> path)
-        {
-            var fieldName = field?.Name;
-
-            var nonNullType = fieldType as NonNullGraphType;
-            if (nonNullType != null)
-            {
-                var type = nonNullType.ResolvedType;
-                var completed = await CompleteValueAsync(context, parentType, type, field, result, path).ConfigureAwait(false);
-                if (completed == null)
-                {
-                    var error = new ExecutionError("Cannot return null for non-null type. Field: {0}, Type: {1}!."
-                        .ToFormat(fieldName, type.Name));
-                    error.AddLocation(field, context.Document);
-                    throw error;
-                }
-
-                return completed;
-            }
-
-            if (result == null)
-            {
-                return null;
-            }
-
-            if (fieldType is ScalarGraphType)
-            {
-                var scalarType = fieldType as ScalarGraphType;
-                var coercedValue = scalarType.Serialize(result);
-                return coercedValue;
-            }
-
-            if (fieldType is ListGraphType)
-            {
-                return await ResolveListFromData(context, result, parentType, fieldType, field, path)
-                    .ConfigureAwait(false);
-            }
-
-            var objectType = fieldType as IObjectGraphType;
-
-            if (fieldType is IAbstractGraphType)
-            {
-                var abstractType = fieldType as IAbstractGraphType;
-                objectType = abstractType.GetObjectType(result);
-
-                if (objectType == null)
-                {
-                    var error = new ExecutionError(
-                        $"Abstract type {abstractType.Name} must resolve to an Object type at " +
-                        $"runtime for field {parentType.Name}.{fieldName} " +
-                        $"with value {result}, received 'null'.");
-                    error.AddLocation(field, context.Document);
-                    throw error;
-                }
-
-                if (!abstractType.IsPossibleType(objectType))
-                {
-                    var error = new ExecutionError(
-                        "Runtime Object type \"{0}\" is not a possible type for \"{1}\""
-                        .ToFormat(objectType, abstractType));
-                    error.AddLocation(field, context.Document);
-                    throw error;
-                }
-            }
-
-            if (objectType == null)
-            {
-                return null;
-            }
-
-            if (objectType.IsTypeOf != null && !objectType.IsTypeOf(result))
-            {
-                var error = new ExecutionError(
-                    "Expected value of type \"{0}\" but got: {1}."
-                    .ToFormat(objectType, result));
-                error.AddLocation(field, context.Document);
-                throw error;
-            }
-
-            var subFields = new Dictionary<string, Field>();
-            var visitedFragments = new List<string>();
-
-            subFields = CollectFields(context, objectType, field?.SelectionSet, subFields, visitedFragments);
-
-            return await ExecuteFieldsAsync(context, objectType, result, subFields, path).ConfigureAwait(false);
-        }
-
-        /// <summary>
-        /// Resolve lists in a performant manner
-        /// </summary>
-        private async Task<IList<object>> ResolveListFromData(ExecutionContext context, object source, IObjectGraphType parentType,
-            IGraphType graphType, Field field, IEnumerable<string> path)
-        {
-            var listInfo = graphType as ListGraphType;
-            var subType = listInfo?.ResolvedType as IObjectGraphType;
-            var data = source as IEnumerable;
-            var visitedFragments = new List<string>();
-            var subFields = CollectFields(context, subType, field.SelectionSet, null, visitedFragments);
-
-            if (data == null)
-            {
-                var error = new ExecutionError("User error: expected an IEnumerable list though did not find one.");
-                error.AddLocation(field, context.Document);
-                throw error;
-            }
-
-            var index = 0;
-            var tasks = new List<Task<object>>();
-
-            foreach (var node in data)
-            {
-                var currentPath = path.Concat(new[] {$"{index++}"});
-
-                if (subType != null)
-                {
-                    var nodeTask = ExecuteFieldsAsync(context, subType, node, subFields, currentPath)
-                        .ContinueWith(task => (object)task.Result);
-
-                    tasks.Add(nodeTask);
-                }
-                else
-                {
-                    var nodeTask = CompleteValueAsync(context, parentType, listInfo?.ResolvedType, field, node, currentPath);
-
-                    tasks.Add(nodeTask);
-                }
-            }
-
-            return await Task.WhenAll(tasks)
-                .ConfigureAwait(false);
-        }
-
-        public Dictionary<string, object> GetArgumentValues(ISchema schema, QueryArguments definitionArguments, Arguments astArguments, Variables variables)
-        {
-            if (definitionArguments == null || !definitionArguments.Any())
-            {
-                return null;
-            }
-
-            return definitionArguments.Aggregate(new Dictionary<string, object>(), (acc, arg) =>
-            {
-                var value = astArguments?.ValueFor(arg.Name);
-                var type = arg.ResolvedType;
-
-                var coercedValue = CoerceValue(schema, type, value, variables);
-                coercedValue = coercedValue ?? arg.DefaultValue;
-                acc[arg.Name] = coercedValue;
-
-                return acc;
-            });
-        }
-
-        public FieldType GetFieldDefinition(Document document, ISchema schema, IObjectGraphType parentType, Field field)
-        {
-            if (field.Name == SchemaIntrospection.SchemaMeta.Name && schema.Query == parentType)
-            {
-                return SchemaIntrospection.SchemaMeta;
-            }
-            if (field.Name == SchemaIntrospection.TypeMeta.Name && schema.Query == parentType)
-            {
-                return SchemaIntrospection.TypeMeta;
-            }
-            if (field.Name == SchemaIntrospection.TypeNameMeta.Name)
-            {
-                return SchemaIntrospection.TypeNameMeta;
-            }
-
-            if (parentType == null)
-            {
-                var error = new ExecutionError($"Schema is not configured correctly to fetch {field.Name}.  Are you missing a root type?");
-                error.AddLocation(field, document);
-                throw error;
-            }
-
-            return parentType.Fields.FirstOrDefault(f => f.Name == field.Name);
         }
 
         public IObjectGraphType GetOperationRootType(Document document, ISchema schema, Operation operation)
@@ -794,336 +337,23 @@ namespace GraphQL
             return type;
         }
 
-        public Variables GetVariableValues(Document document, ISchema schema, VariableDefinitions variableDefinitions, Inputs inputs)
+        protected virtual IExecutionStrategy SelectExecutionStrategy(ExecutionContext context)
         {
-            var variables = new Variables();
-            variableDefinitions?.Apply(v =>
+            // TODO: Should we use cached instances of the default execution strategies?
+            switch (context.Operation.OperationType)
             {
-                var variable = new Variable();
-                variable.Name = v.Name;
+                case OperationType.Query:
+                    return new ParallelExecutionStrategy();
 
-                object variableValue = null;
-                inputs?.TryGetValue(v.Name, out variableValue);
-                variable.Value = GetVariableValue(document, schema, v, variableValue);
+                case OperationType.Mutation:
+                    return new SerialExecutionStrategy();
 
-                variables.Add(variable);
-            });
-            return variables;
-        }
+                case OperationType.Subscription:
+                    return new SubscriptionExecutionStrategy();
 
-        public object GetVariableValue(Document document, ISchema schema, VariableDefinition variable, object input)
-        {
-            var type = variable.Type.GraphTypeFromType(schema);
-
-            try
-            {
-                AssertValidValue(schema, type, input, variable.Name);
+                default:
+                    throw new InvalidOperationException($"Unexpected OperationType {context.Operation.OperationType}");
             }
-            catch (InvalidValueException error)
-            {
-                error.AddLocation(variable, document);
-                throw;
-            }
-
-            if (input == null)
-            {
-                if (variable.DefaultValue != null)
-                {
-                    return variable.DefaultValue.Value;
-                }
-            }
-            var coercedValue = CoerceValue(schema, type, input.AstFromValue(schema, type));
-            return coercedValue;
-        }
-
-        public void AssertValidValue(ISchema schema, IGraphType type, object input, string fieldName)
-        {
-            if (type is NonNullGraphType)
-            {
-                var nonNullType = ((NonNullGraphType)type).ResolvedType;
-
-                if (input == null)
-                {
-                    throw new InvalidValueException(fieldName, "Received a null input for a non-null field.");
-                }
-
-                AssertValidValue(schema, nonNullType, input, fieldName);
-                return;
-            }
-
-            if (input == null)
-            {
-                return;
-            }
-
-            if (type is ScalarGraphType)
-            {
-                var scalar = (ScalarGraphType)type;
-                if (ValueFromScalar(scalar, input) == null)
-                    throw new InvalidValueException(fieldName, "Invalid Scalar value for input field.");
-
-                return;
-            }
-
-            if (type is ListGraphType)
-            {
-                var listType = (ListGraphType)type;
-                var listItemType = listType.ResolvedType;
-
-                var list = input as IEnumerable;
-                if (list != null && !(input is string))
-                {
-                    var index = -1;
-                    foreach (var item in list)
-                        AssertValidValue(schema, listItemType, item, $"{fieldName}[{++index}]");
-                }
-                else
-                {
-                    AssertValidValue(schema, listItemType, input, fieldName);
-                }
-                return;
-            }
-
-            if (type is IObjectGraphType || type is IInputObjectGraphType)
-            {
-                var dict = input as Dictionary<string, object>;
-                var complexType = (IComplexGraphType)type;
-
-                if (dict == null)
-                {
-                    throw new InvalidValueException(fieldName,
-                        $"Unable to parse input as a '{type.Name}' type. Did you provide a List or Scalar value accidentally?");
-                }
-
-                // ensure every provided field is defined
-                var unknownFields = type is IInputObjectGraphType
-                    ? dict.Keys.Where(key => complexType.Fields.All(field => field.Name != key)).ToArray()
-                    : null;
-
-                if (unknownFields != null && unknownFields.Any())
-                {
-                    throw new InvalidValueException(fieldName,
-                        $"Unrecognized input fields {string.Join(", ", unknownFields.Select(k => $"'{k}'"))} for type '{type.Name}'.");
-                }
-
-                foreach (var field in complexType.Fields)
-                {
-                    object fieldValue;
-                    dict.TryGetValue(field.Name, out fieldValue);
-                    AssertValidValue(schema, field.ResolvedType, fieldValue, $"{fieldName}.{field.Name}");
-                }
-                return;
-            }
-
-            throw new InvalidValueException(fieldName ?? "input", "Invalid input");
-        }
-
-        private object ValueFromScalar(ScalarGraphType scalar, object input)
-        {
-            if (input is IValue)
-            {
-                return scalar.ParseLiteral((IValue)input);
-            }
-
-            return scalar.ParseValue(input);
-        }
-
-        public object CoerceValue(ISchema schema, IGraphType type, IValue input, Variables variables = null)
-        {
-            if (type is NonNullGraphType)
-            {
-                var nonNull = type as NonNullGraphType;
-                return CoerceValue(schema, nonNull.ResolvedType, input, variables);
-            }
-
-            if (input == null)
-            {
-                return null;
-            }
-
-            var variable = input as VariableReference;
-            if (variable != null)
-            {
-                return variables?.ValueFor(variable.Name);
-            }
-
-            if (type is ListGraphType)
-            {
-                var listType = type as ListGraphType;
-                var listItemType = listType.ResolvedType;
-                var list = input as ListValue;
-                return list != null
-                    ? list.Values.Map(item => CoerceValue(schema, listItemType, item, variables)).ToArray()
-                    : new[] { CoerceValue(schema, listItemType, input, variables) };
-            }
-
-            if (type is IObjectGraphType || type is IInputObjectGraphType)
-            {
-                var complexType = type as IComplexGraphType;
-                var obj = new Dictionary<string, object>();
-
-                var objectValue = input as ObjectValue;
-                if (objectValue == null)
-                {
-                    return null;
-                }
-
-                complexType.Fields.Apply(field =>
-                {
-                    var objectField = objectValue.Field(field.Name);
-                    if (objectField != null)
-                    {
-                        var fieldValue = CoerceValue(schema, field.ResolvedType, objectField.Value, variables);
-                        fieldValue = fieldValue ?? field.DefaultValue;
-
-                        obj[field.Name] = fieldValue;
-                    }
-                });
-
-                return obj;
-            }
-
-            if (type is ScalarGraphType)
-            {
-                var scalarType = type as ScalarGraphType;
-                return scalarType.ParseLiteral(input);
-            }
-
-            return null;
-        }
-
-        public Dictionary<string, Field> CollectFields(
-            ExecutionContext context,
-            IGraphType specificType,
-            SelectionSet selectionSet,
-            Dictionary<string, Field> fields,
-            List<string> visitedFragmentNames)
-        {
-            if (fields == null)
-            {
-                fields = new Dictionary<string, Field>();
-            }
-
-            selectionSet?.Selections.Apply(selection =>
-            {
-                if (selection is Field)
-                {
-                    var field = (Field)selection;
-                    if (!ShouldIncludeNode(context, field.Directives))
-                    {
-                        return;
-                    }
-
-                    var name = field.Alias ?? field.Name;
-                    fields[name] = field;
-                }
-                else if (selection is FragmentSpread)
-                {
-                    var spread = (FragmentSpread)selection;
-
-                    if (visitedFragmentNames.Contains(spread.Name)
-                        || !ShouldIncludeNode(context, spread.Directives))
-                    {
-                        return;
-                    }
-
-                    visitedFragmentNames.Add(spread.Name);
-
-                    var fragment = context.Fragments.FindDefinition(spread.Name);
-                    if (fragment == null
-                        || !ShouldIncludeNode(context, fragment.Directives)
-                        || !DoesFragmentConditionMatch(context, fragment.Type.Name, specificType))
-                    {
-                        return;
-                    }
-
-                    CollectFields(context, specificType, fragment.SelectionSet, fields, visitedFragmentNames);
-                }
-                else if (selection is InlineFragment)
-                {
-                    var inline = (InlineFragment)selection;
-
-                    var name = inline.Type != null ? inline.Type.Name : specificType.Name;
-
-                    if (!ShouldIncludeNode(context, inline.Directives)
-                      || !DoesFragmentConditionMatch(context, name, specificType))
-                    {
-                        return;
-                    }
-
-                    CollectFields(context, specificType, inline.SelectionSet, fields, visitedFragmentNames);
-                }
-            });
-
-            return fields;
-        }
-
-        public bool ShouldIncludeNode(ExecutionContext context, Directives directives)
-        {
-            if (directives != null)
-            {
-                var directive = directives.Find(DirectiveGraphType.Skip.Name);
-                if (directive != null)
-                {
-                    var values = GetArgumentValues(
-                        context.Schema,
-                        DirectiveGraphType.Skip.Arguments,
-                        directive.Arguments,
-                        context.Variables);
-
-                    object ifObj;
-                    values.TryGetValue("if", out ifObj);
-
-                    bool ifVal;
-                    return !(bool.TryParse(ifObj?.ToString() ?? string.Empty, out ifVal) && ifVal);
-                }
-
-                directive = directives.Find(DirectiveGraphType.Include.Name);
-                if (directive != null)
-                {
-                    var values = GetArgumentValues(
-                        context.Schema,
-                        DirectiveGraphType.Include.Arguments,
-                        directive.Arguments,
-                        context.Variables);
-
-                    object ifObj;
-                    values.TryGetValue("if", out ifObj);
-
-                    bool ifVal;
-                    return bool.TryParse(ifObj?.ToString() ?? string.Empty, out ifVal) && ifVal;
-                }
-            }
-
-            return true;
-        }
-
-        public bool DoesFragmentConditionMatch(ExecutionContext context, string fragmentName, IGraphType type)
-        {
-            if (string.IsNullOrWhiteSpace(fragmentName))
-            {
-                return true;
-            }
-
-            var conditionalType = context.Schema.FindType(fragmentName);
-
-            if (conditionalType == null)
-            {
-                return false;
-            }
-
-            if (conditionalType.Equals(type))
-            {
-                return true;
-            }
-
-            if (conditionalType is IAbstractGraphType)
-            {
-                var abstractType = (IAbstractGraphType)conditionalType;
-                return abstractType.IsPossibleType(type);
-            }
-
-            return false;
         }
     }
 }
